build: Fix dependency loop cause by phony targets
When building proj-deps.mk, it would always be evaluated as
'out-of-date' because the dependency collection relied on
phony targets.

And since make 'reloads' itself after including a file it
generates, it would continuously do so ... at least on some
version of make. make >= 4.2.1 seem to not have this issue.

To solve this, move on to real file 'markers' to indicate
the dependency has been processed. Also make sure it depends
on core.mk so it rebuilds it the core config changes

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>